create "shape_pager" liquid tag, remove "set_properties" liquid tag
