+++ src/OrchardCore.Modules/OrchardCore.Lists/Drivers/ListPartDisplayDriver.cs
@@ -74,7 +74,6 @@ namespace OrchardCore.Lists.Drivers
                 var query = _session.Query<ContentItem>()
                     .With<ContainedPartIndex>(x => x.ListContentItemId == listPart.ContentItem.ContentItemId)
                     .With<ContentItemIndex>(CreateContentIndexFilter(beforeValue, null, published))
-                    .OrderBy(x => x.CreatedUtc)
                     .Take(pager.PageSize + 1);
 
                 var containedItems = await query.ListAsync();
@@ -84,9 +83,9 @@ namespace OrchardCore.Lists.Drivers
                     return containedItems;
                 }
 
-                containedItems = containedItems.Reverse();
+                containedItems = containedItems.OrderByDescending(x => x.CreatedUtc);
 
-                // There is always an After ras we clicked on Before
+                // There is always an After as we clicked on Before
                 pager.Before = null;
                 pager.After = containedItems.Last().CreatedUtc.Value.Ticks.ToString();
 
@@ -104,7 +103,6 @@ namespace OrchardCore.Lists.Drivers
                 var query = _session.Query<ContentItem>()
                     .With<ContainedPartIndex>(x => x.ListContentItemId == listPart.ContentItem.ContentItemId)
                     .With<ContentItemIndex>(CreateContentIndexFilter(null, afterValue, published))
-                    .OrderByDescending(x => x.CreatedUtc)
                     .Take(pager.PageSize + 1);
 
                 var containedItems = await query.ListAsync();
@@ -113,6 +111,8 @@ namespace OrchardCore.Lists.Drivers
                 {
                     return containedItems;
                 }
+
+                containedItems = containedItems.OrderByDescending(x => x.CreatedUtc);
 
                 // There is always a Before page as we clicked on After
                 pager.Before = containedItems.First().CreatedUtc.Value.Ticks.ToString();
@@ -128,11 +128,22 @@ namespace OrchardCore.Lists.Drivers
             }
             else
             {
+                var count = await _session.Query<ContentItem>()
+                    .With<ContainedPartIndex>(x => x.ListContentItemId == listPart.ContentItem.ContentItemId)
+                    .With<ContentItemIndex>(x => x.Published)
+                    .CountAsync();
+
+                var skip = 0;
+
+                if (count > pager.PageSize + 1)
+                {
+                    skip = count - (pager.PageSize + 1);
+                }
+
                 var query = _session.Query<ContentItem>()
                     .With<ContainedPartIndex>(x => x.ListContentItemId == listPart.ContentItem.ContentItemId)
                     .With<ContentItemIndex>(CreateContentIndexFilter(null, null, published))
-                    .OrderByDescending(x => x.CreatedUtc)
-                    .Take(pager.PageSize + 1);
+                    .Skip(skip);
 
                 var containedItems = await query.ListAsync();
 
@@ -140,6 +151,8 @@ namespace OrchardCore.Lists.Drivers
                 {
                     return containedItems;
                 }
+
+                containedItems = containedItems.OrderByDescending(x => x.CreatedUtc);
 
                 pager.Before = null;
                 pager.After = null;
